requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
cycler==0.11.0
debuggerModule==0.4
Flask==0.12.2
Flask-Login==0.5.0
Flask-RESTful==0.3.6
fonttools==4.29.1
gunicorn==19.7.1
idna==2.5
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
more-itertools==8.12.0
multitasking==0.0.10
Nasdaq-Data-Link==1.0.1
numpy==1.22.2
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
requests2==2.16.0
six==1.16.0
soupsieve==2.3.1
urllib3==1.21.1
Werkzeug==2.0.2
yahoofinancials==1.6
yfinance==0.1.70
